frontend/requirements.txt
streamlit
requests
pytesseract
opencv-python
pillow
